feat(destroying): add health-damage system
tremendously insane (!!! \(2o0)/ !!!) health-damage system for pigs and
fortification objects. There are an abstract class with common damage
detection / applying logic, so you can't put it on object in the inspector
accidentally, and derivative classes for pigs and fortifications
respectively for inserting specific logic on destroying, such as
notification about the defeated pig - key element for passing the level
(very clever system (!!!) wow (!!!)).